requirements.txt
fastapi
uvicorn
websockets
langchain
langchain-openai
langchain-community
chromadb
deepgram-sdk==2.12.0
google-cloud-texttospeech
twilio
langgraph
python-dotenv
gunicorn